% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocations_by_indexer.R
\name{allocations_by_indexer}
\alias{allocations_by_indexer}
\title{Function to acquire allocation data from an Indexer.}
\usage{
allocations_by_indexer(wallet_address, my_api_key)
}
\arguments{
\item{wallet_address}{Wallet address of Indexer to obtain allocation data from.}

\item{my_api_key}{Add your API key.}
}
\value{
Returns allocations from specified Indexer.
}
\description{
Function to acquire allocation data from an Indexer.
}
