% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approve_actions.R
\name{approve_actions}
\alias{approve_actions}
\title{Generate approve action command for Indexer allocations.}
\usage{
approve_actions(actions)
}
\arguments{
\item{actions}{Add numbers with first and last total separated by colon (:).}
}
\value{
A string to approve Indexer allocation actions.
}
\description{
Generate approve action command for Indexer allocations.
}
\examples{
#graph indexer actions queue allocate deployment_id grt_amount --network arbitrum-one
approve_actions(1:4)

}
