% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_epoch.R
\name{current_epoch}
\alias{current_epoch}
\title{Get the current epoch for The Graph protocol.}
\usage{
current_epoch(my_api_key)
}
\arguments{
\item{my_api_key}{Add your API key.}
}
\value{
Returns the current epoch for The Graph protocol.
}
\description{
Get the current epoch for The Graph protocol.
}
\examples{
the_epoch = current_epoch("3a2f2d6ce1672801c3a044139373a287")
the_epoch

}
