% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_subgraph.R
\name{query_subgraph}
\alias{query_subgraph}
\title{Returns subgraph data from an indexed smart contract on The Graph Network.}
\usage{
query_subgraph(graphql_query, subgraph_id, api_key)
}
\arguments{
\item{graphql_query}{A GraphQL query of a subgraph.}

\item{subgraph_id}{Subgraph ID (found near top of subgraph page on Graph Explorer)}

\item{api_key}{An API key from Subgraph Studio.}
}
\value{
Returns subgraph data from an indexed smart contract.
}
\description{
Returns subgraph data from an indexed smart contract on The Graph Network.
}
