% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraph_display_names.R
\name{subgraph_display_names}
\alias{subgraph_display_names}
\title{Function returns subgraph display names.}
\usage{
subgraph_display_names(skip, first, my_api_key)
}
\arguments{
\item{skip}{Number of records to skip in GraphQL query records.}

\item{first}{Number of records to return in GraphQL query records.}

\item{my_api_key}{Add your API key.}
}
\value{
Returns subgraph display names.
}
\description{
Function returns subgraph display names.
}
