% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp_to_unix.R
\name{timestamp_to_unix}
\alias{timestamp_to_unix}
\title{Get the unix from a timestamp.}
\usage{
timestamp_to_unix(timestamp)
}
\arguments{
\item{timestamp}{Add a timestamp.}
}
\value{
Returns the unix from a timestamp.
}
\description{
Get the unix from a timestamp.
}
\examples{
timestamp = "2024-09-14"
the_unix = timestamp_to_unix(timestamp)
the_unix

}
